requirements.txt
fastapi==0.128.0
openai==2.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.0.1
requests==2.32.5
uvicorn==0.39.0
